refactor: extract Analytics to a separate file
