+++ src/contexts/AnalyticsProviderContext.ts
@@ -1,16 +1,7 @@
 import {createContext} from 'react';
-import {UnknownRecord} from '../types/common';
+import {Analytics} from '../core';
 
-export interface Analytics {
-  onPageView(params?: UnknownRecord): void;
-  onEvent(name: string, params?: UnknownRecord): void;
-  onClick(name: string, params?: UnknownRecord): void;
-  onSet(...args: [string, UnknownRecord] | [UnknownRecord]): void;
-  onSetUserId(userId: string | null): void;
-  onSetUserProperty(params: UnknownRecord): void;
-}
-
-export const initialState: Analytics = {
+const initialState: Analytics = {
   onPageView: () => null,
   onEvent: () => null,
   onClick: () => null,
