+++ src/index.tsx
@@ -2,3 +2,4 @@ export * from './components';
 export * from './utils';
 export * from './contexts';
 export * from './hooks';
+export * from './core';
